requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib
python-multipart
jinja2
alembic
email-validator
bcrypt==4.0.1
pytest
pytest-asyncio
pytest-dotenv
httpx
Faker